Move project.toml parsing out of internal package
This make it possible for this code to be used as 
a library.

Fixes #657
