+++ internal/commands/build.go
@@ -15,9 +15,9 @@ import (
 	"github.com/buildpacks/pack"
 	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/paths"
-	"github.com/buildpacks/pack/internal/project"
 	"github.com/buildpacks/pack/internal/style"
 	"github.com/buildpacks/pack/logging"
+	"github.com/buildpacks/pack/project"
 )
 
 type BuildFlags struct {
